requirements.txt
asgiref==3.2.10
Django==3.1
djangorestframework==3.14.0
Pillow==9.5.0
psycopg2-binary==2.9.5
pytz==2023.3
sqlparse==0.4.3
